Relocating Tests & Updating references
Moving the pattern validation test constructs into a sub directory
intended to hold regex-type testing elements.
